+++ src/cs5004/animator/model/shape/Dimension.java
@@ -5,48 +5,48 @@ package cs5004.animator.model.shape;
  * through width and height values.
  */
 public class Dimension {
-  protected double width;
-  protected double height;
+  protected int width;
+  protected int height;
 
   /**
    * Dimensions constructor takes two doubles for width and height of the shape.
-   * @param width width of the shape, double
-   * @param height height of the shape, double
+   * @param width width of the shape, int
+   * @param height height of the shape, int
    */
-  public Dimension(double width, double height) {
+  public Dimension(int width, int height) {
     this.width = width;
     this.height = height;
   }
 
   /**
    * Get the shape's width.
-   * @return width, double
+   * @return width, int
    */
-  public double getWidth() {
+  public int getWidth() {
     return this.width;
   }
 
   /**
    * Get the shape's height.
-   * @return height, double
+   * @return height, int
    */
-  public double getHeight() {
+  public int getHeight() {
     return this.height;
   }
 
   /**
    * Setter method for width of the shape.
-   * @param sWidth width being set as the new width for the shape, double
+   * @param sWidth width being set as the new width for the shape, int
    */
-  public void setWidth(double sWidth) {
+  public void setWidth(int sWidth) {
     this.width = sWidth;
   }
 
   /**
    * Setter method for height of the shape.
-   * @param sHeight height being set as the new height for the shape, double
+   * @param sHeight height being set as the new height for the shape, int
    */
-  public void setHeight(double sHeight) {
+  public void setHeight(int sHeight) {
     this.height = sHeight;
   }
 }
